Allow merge to be used with git mergetool
E.g. when using git rebase the normal git-merge tool does not seem to be
used. Re-use the existing mergers and deal with the different formats. I am
not fully sure it works all right but on first merge it seemed to have done the
right thing (even if version didn't change).

~/.gitconfig:
[mergetool "mcmerge"]
        cmd = /GitFileTree-MergeDriver/merge --mergetool $BASE $LOCAL $REMOTE $MERGED

git mergetool --tool=mcmerge